+++ extensions/native/circuit/tests/integration_test.rs
@@ -16,6 +16,8 @@ use openvm_circuit::{
     system::{memory::CHUNK, program::trace::VmCommittedExe, SystemCpuBuilder},
     utils::{air_test, air_test_with_min_segments, test_system_config_without_continuations},
 };
+#[cfg(feature = "cuda")]
+use openvm_cuda_backend::engine::GpuBabyBearPoseidon2Engine as TestEngine;
 use openvm_instructions::{
     exe::VmExe,
     instruction::Instruction,
@@ -25,9 +27,13 @@ use openvm_instructions::{
     SysPhantom,
     SystemOpcode::*,
 };
+#[cfg(not(feature = "cuda"))]
+use openvm_native_circuit::NativeCpuBuilder as NativeBuilder;
+#[cfg(feature = "cuda")]
+use openvm_native_circuit::NativeGpuBuilder as NativeBuilder;
 use openvm_native_circuit::{
     execute_program, test_native_config, test_native_continuations_config,
-    test_rv32_with_kernels_config, NativeConfig, NativeCpuBuilder,
+    test_rv32_with_kernels_config, NativeConfig,
 };
 use openvm_native_compiler::{
     CastfOpcode,
@@ -42,11 +48,13 @@ use openvm_rv32im_transpiler::BranchEqualOpcode::*;
 use openvm_stark_backend::{
     config::StarkGenericConfig, engine::StarkEngine, p3_field::FieldAlgebra,
 };
+#[cfg(not(feature = "cuda"))]
+use openvm_stark_sdk::config::baby_bear_poseidon2::BabyBearPoseidon2Engine as TestEngine;
 use openvm_stark_sdk::{
     config::{
-        baby_bear_poseidon2::{BabyBearPoseidon2Config, BabyBearPoseidon2Engine},
-        fri_params::standard_fri_params_with_100_bits_conjectured_security,
-        setup_tracing, FriParameters,
+        baby_bear_poseidon2::BabyBearPoseidon2Config,
+        fri_params::standard_fri_params_with_100_bits_conjectured_security, setup_tracing,
+        FriParameters,
     },
     engine::StarkFriEngine,
     p3_baby_bear::BabyBear,
@@ -102,13 +110,13 @@ fn test_vm_1() {
 
     let program = Program::from_instructions(&instructions);
 
-    air_test(NativeCpuBuilder, test_native_config(), program);
+    air_test(NativeBuilder, test_native_config(), program);
 }
 
 // See crates/sdk/src/prover/root.rs for intended usage
 #[test]
 fn test_vm_override_trace_heights() -> eyre::Result<()> {
-    let e = BabyBearPoseidon2Engine::new(FriParameters::standard_fast());
+    let e = TestEngine::new(FriParameters::standard_fast());
     let program = Program::<BabyBear>::from_instructions(&[
         Instruction::large_from_isize(ADD.global_opcode(), 0, 4, 0, 4, 0, 0, 0),
         Instruction::from_isize(TERMINATE.global_opcode(), 0, 0, 0, 0, 0),
@@ -125,7 +133,7 @@ fn test_vm_override_trace_heights() -> eyre::Result<()> {
 
     // Test getting heights.
     let vm_config = NativeConfig::aggregation(8, 3);
-    let (mut vm, pk) = VirtualMachine::new_with_keygen(e, NativeCpuBuilder, vm_config)?;
+    let (mut vm, pk) = VirtualMachine::new_with_keygen(e, NativeBuilder, vm_config)?;
     let vk = pk.get_vk();
 
     let state = vm.create_initial_state(&committed_exe.exe, vec![]);
@@ -169,9 +177,8 @@ fn test_vm_1_optional_air() -> eyre::Result<()> {
     // Aggregation VmConfig has Core/Poseidon2/FieldArithmetic/FieldExtension chips. The program
     // only uses Core and FieldArithmetic. All other chips should not have AIR proof inputs.
     let config = NativeConfig::aggregation(4, 3);
-    let engine =
-        BabyBearPoseidon2Engine::new(standard_fri_params_with_100_bits_conjectured_security(3));
-    let (vm, pk) = VirtualMachine::new_with_keygen(engine, NativeCpuBuilder, config)?;
+    let engine = TestEngine::new(standard_fri_params_with_100_bits_conjectured_security(3));
+    let (vm, pk) = VirtualMachine::new_with_keygen(engine, NativeBuilder, config)?;
     let num_airs = pk.per_air.len();
 
     let n = 6;
@@ -205,8 +212,7 @@ fn test_vm_public_values() -> eyre::Result<()> {
     let num_public_values = 100;
     let config = test_system_config_without_continuations().with_public_values(num_public_values);
     assert!(!config.continuation_enabled);
-    let engine =
-        BabyBearPoseidon2Engine::new(standard_fri_params_with_100_bits_conjectured_security(3));
+    let engine = TestEngine::new(standard_fri_params_with_100_bits_conjectured_security(3));
     let (vm, pk) = VirtualMachine::new_with_keygen(engine, SystemCpuBuilder, config)?;
 
     let instructions = vec![
@@ -271,18 +277,18 @@ fn test_vm_initial_memory() {
         init_memory,
         fn_bounds: Default::default(),
     };
-    air_test(NativeCpuBuilder, config, exe);
+    air_test(NativeBuilder, config, exe);
 }
 
 #[test]
 fn test_vm_1_persistent() -> eyre::Result<()> {
-    let engine = BabyBearPoseidon2Engine::new(FriParameters::standard_fast());
+    let engine = TestEngine::new(FriParameters::standard_fast());
     let config = test_native_continuations_config();
     let merkle_air_idx = config.system.memory_boundary_air_id() + 1;
     let ptr_max_bits = config.system.memory_config.pointer_max_bits;
     let addr_space_height = config.system.memory_config.addr_space_height;
 
-    let (vm, pk) = VirtualMachine::new_with_keygen(engine, NativeCpuBuilder, config)?;
+    let (vm, pk) = VirtualMachine::new_with_keygen(engine, NativeBuilder, config)?;
 
     let n = 6;
     let instructions = vec![
@@ -375,7 +381,7 @@ fn test_vm_without_field_arithmetic() {
 
     let program = Program::from_instructions(&instructions);
 
-    air_test(NativeCpuBuilder, test_native_config(), program);
+    air_test(NativeBuilder, test_native_config(), program);
 }
 
 #[test]
@@ -422,7 +428,7 @@ fn test_vm_fibonacci_old() {
 
     let program = Program::from_instructions(&instructions);
 
-    air_test(NativeCpuBuilder, test_native_config(), program);
+    air_test(NativeBuilder, test_native_config(), program);
 }
 
 #[test]
@@ -481,7 +487,7 @@ fn test_vm_fibonacci_old_cycle_tracker() {
 
     let program = Program::from_instructions(&instructions);
 
-    air_test(NativeCpuBuilder, test_native_config(), program);
+    air_test(NativeBuilder, test_native_config(), program);
 }
 
 #[test]
@@ -505,7 +511,7 @@ fn test_vm_field_extension_arithmetic() {
 
     let program = Program::from_instructions(&instructions);
 
-    air_test(NativeCpuBuilder, test_native_config(), program);
+    air_test(NativeBuilder, test_native_config(), program);
 }
 
 #[test]
@@ -554,7 +560,7 @@ fn test_vm_max_access_adapter_8() {
             num_sys_airs2 + num_ext_airs
         );
     }
-    air_test(NativeCpuBuilder, test_native_config(), program);
+    air_test(NativeBuilder, test_native_config(), program);
 }
 
 #[test]
@@ -578,7 +584,7 @@ fn test_vm_field_extension_arithmetic_persistent() {
 
     let program = Program::from_instructions(&instructions);
     let config = test_native_continuations_config();
-    air_test(NativeCpuBuilder, config, program);
+    air_test(NativeBuilder, config, program);
 }
 
 #[test]
@@ -639,7 +645,7 @@ fn test_vm_hint() {
 
     let input_stream: Vec<Vec<F>> = vec![vec![F::TWO]];
     let config = test_native_config();
-    air_test_with_min_segments(NativeCpuBuilder, config, program, input_stream, 1);
+    air_test_with_min_segments(NativeBuilder, config, program, input_stream, 1);
 }
 
 #[test]
@@ -891,8 +897,8 @@ fn test_single_segment_executor_no_segmentation() {
         .system
         .set_segmentation_limits(SegmentationLimits::default().with_max_trace_height(1));
 
-    let engine = BabyBearPoseidon2Engine::new(FriParameters::new_for_testing(3));
-    let (vm, _) = VirtualMachine::new_with_keygen(engine, NativeCpuBuilder, config).unwrap();
+    let engine = TestEngine::new(FriParameters::new_for_testing(3));
+    let (vm, _) = VirtualMachine::new_with_keygen(engine, NativeBuilder, config).unwrap();
     let instructions: Vec<_> = (0..2 * DEFAULT_SEGMENT_CHECK_INSNS)
         .map(|_| Instruction::large_from_isize(ADD.global_opcode(), 0, 0, 1, 4, 0, 0, 0))
         .chain(std::iter::once(Instruction::from_isize(
@@ -922,8 +928,8 @@ fn test_vm_execute_metered_cost_native_chips() {
     setup_tracing();
     let config = test_native_config();
 
-    let engine = BabyBearPoseidon2Engine::new(FriParameters::new_for_testing(3));
-    let (vm, _) = VirtualMachine::new_with_keygen(engine, NativeCpuBuilder, config).unwrap();
+    let engine = TestEngine::new(FriParameters::new_for_testing(3));
+    let (vm, _) = VirtualMachine::new_with_keygen(engine, NativeBuilder, config).unwrap();
 
     let instructions = vec![
         // Field Arithmetic operations (FieldArithmeticChip)
@@ -958,9 +964,8 @@ fn test_vm_execute_metered_cost_halt() {
     setup_tracing();
     let config = test_native_config();
 
-    let engine = BabyBearPoseidon2Engine::new(FriParameters::new_for_testing(3));
-    let (vm, _) =
-        VirtualMachine::new_with_keygen(engine, NativeCpuBuilder, config.clone()).unwrap();
+    let engine = TestEngine::new(FriParameters::new_for_testing(3));
+    let (vm, _) = VirtualMachine::new_with_keygen(engine, NativeBuilder, config.clone()).unwrap();
 
     let instructions = vec![
         // Field Arithmetic operations (FieldArithmeticChip)
